Updated the Number trait
